audioservice: unified socket interfaces for cast lite transport [1/1]
PD#SWPL-131887

Problem:
1.the server-side program currently in use is based on c++, and needs to be
changed to be based on c for the sake of generality.
2.cast lite based on socket communication between client and server using the
interface is still mainly avs interface, which is not a common interface, so
it should be abstracted into a common interface;
3.lack of server-side demo
4.the original server logic is unable to mass message all clients that
register for the same type of event (cause the map that registers the
message can only be associated with one client at a time)
5.solve the problem of program crash caused by duplicate connect in the
original client code;
6.dynamic libraries add aml log printing.

Solution:
1.rewrite the server related code into c and also write a server and client
test set.
2.the relevant interfaces are updated from the avs directory to the /amlogic/
aml_commonlib directory to further abstract the relevant interfaces;
3.assign a bitmap to each client socket to solve problem 4;
4.using a mutex lock on the thread that calls the connect initialization action

Verify:
A5.

Change-Id: I449d5362571a34a34914f9e003b71b4f9dc482de
Signed-off-by: yifan.li <yifan.li@amlogic.com>